adding the documentation to all the enhanced modview models
