+++ app/src/main/java/com/example/clicker/presentation/enhancedModView/views/ModVersionThree.kt
@@ -1,4 +1,4 @@
-package com.example.clicker.presentation.enhancedModView
+package com.example.clicker.presentation.enhancedModView.views
 
 import android.util.Log
 import androidx.compose.foundation.BorderStroke
@@ -35,9 +35,6 @@ import androidx.compose.foundation.text.appendInlineContent
 import androidx.compose.foundation.text.selection.LocalTextSelectionColors
 import androidx.compose.foundation.text.selection.TextSelectionColors
 import androidx.compose.material.CircularProgressIndicator
-import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.ModalBottomSheetLayout
-import androidx.compose.material.ModalBottomSheetValue
 import androidx.compose.material.TextField
 import androidx.compose.material.TextFieldDefaults
 
@@ -47,7 +44,6 @@ import androidx.compose.material.icons.filled.Check
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.Menu
 import androidx.compose.material.icons.filled.Person
-import androidx.compose.material.rememberModalBottomSheetState
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.CardDefaults
@@ -84,7 +80,6 @@ import androidx.compose.ui.platform.LocalHapticFeedback
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.TextRange
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.withStyle
@@ -100,31 +95,22 @@ import com.example.clicker.network.models.websockets.TwitchUserData
 import com.example.clicker.network.repository.ClickedUnbanRequestInfo
 
 import com.example.clicker.network.repository.models.EmoteListMap
-import com.example.clicker.network.repository.models.EmoteNameUrl
 import com.example.clicker.network.repository.util.AutoModQueueMessage
-import com.example.clicker.presentation.home.disableClickAndRipple
-import com.example.clicker.presentation.modView.AutoModMessageListImmutableCollection
-import com.example.clicker.presentation.modView.ListTitleValue
-import com.example.clicker.presentation.modView.ModActionData
-import com.example.clicker.presentation.modView.ModActionListImmutableCollection
-import com.example.clicker.presentation.modView.ModViewViewModel
-import com.example.clicker.presentation.modView.UnbanRequestItemImmutableCollection
-import com.example.clicker.presentation.stream.StreamViewModel
-import com.example.clicker.presentation.stream.views.chat.chatSettings.ChatSettingsColumn
+import com.example.clicker.presentation.enhancedModView.AutoModMessageListImmutableCollection
+import com.example.clicker.presentation.enhancedModView.ModActionData
+import com.example.clicker.presentation.enhancedModView.ModActionListImmutableCollection
+import com.example.clicker.presentation.enhancedModView.Sections
+import com.example.clicker.presentation.enhancedModView.UnbanRequestItemImmutableCollection
+
 import com.example.clicker.presentation.stream.views.chat.DualIconsButton
-import com.example.clicker.presentation.stream.views.chat.FullChatModView
 import com.example.clicker.presentation.stream.views.chat.ImprovedChatUI
 import com.example.clicker.presentation.stream.views.chat.isScrolledToEnd
 import kotlinx.coroutines.launch
 import kotlin.math.roundToInt
-import com.example.clicker.presentation.modView.slowModeListImmutable
-import com.example.clicker.presentation.modView.followerModeListImmutable
 import com.example.clicker.presentation.sharedViews.SwitchWithIcon
 import com.example.clicker.presentation.stream.models.ClickedUsernameChatsWithDateSentImmutable
-import com.example.clicker.presentation.stream.views.TestingNewBottomModal
 import com.example.clicker.util.Response
 import com.example.clicker.util.WebSocketResponse
-import com.example.clicker.presentation.stream.views.chat.chatSettings.ChatSettingsViewModel
 import com.example.clicker.util.UnAuthorizedResponse
 
 
@@ -134,660 +120,7 @@ import com.example.clicker.util.UnAuthorizedResponse
  * - **ModViewComponentVersionThree**
  *
  * */
-@OptIn(ExperimentalMaterialApi::class)
-@Composable
-fun ModViewComponentVersionThree(
-    closeModView:()->Unit,
-    twitchUserChat: List<TwitchUserData>, // unstable this is the only one I am going to fix
-    streamViewModel: StreamViewModel,// unstable
-    modViewViewModel: ModViewViewModel,// unstable
-    hideSoftKeyboard:()->Unit,
-    modVersionThreeViewModel: ModVersionThreeViewModel,// unstable
-    chatSettingsViewModel: ChatSettingsViewModel,
-){
 
-    val scope = rememberCoroutineScope()
-
-    val clickedChatterModalState = rememberModalBottomSheetState(
-        initialValue = ModalBottomSheetValue.Hidden,
-        skipHalfExpanded = true
-    )
- 
-
-    val unbanRequestModalState= rememberModalBottomSheetState(
-        initialValue = ModalBottomSheetValue.Hidden,
-        skipHalfExpanded = true
-    )
-    val showUnbanRequestBottomModal:()->Unit =remember(unbanRequestModalState) { {
-        scope.launch {
-            unbanRequestModalState.show()
-        }
-    } }
-
-    val chatSettingsModalState = rememberModalBottomSheetState(
-        initialValue = ModalBottomSheetValue.Hidden,
-        skipHalfExpanded = true
-    )
-
-    val updateClickedUser:(String,String,Boolean,Boolean)->Unit = remember(streamViewModel) { { username, userId, banned, isMod ->
-        streamViewModel.updateClickedChat(
-            username,
-            userId,
-            banned,
-            isMod
-        )
-    } }
-    val doubleClickChat:(String)->Unit =remember(streamViewModel) { {
-        streamViewModel.sendDoubleTapEmote(it)
-    } }
-
-    /******BOX-ONE FUNCTIONS*******/
-    val setBoxOneDoubleTap:(newValue:Boolean)->Unit = remember(modVersionThreeViewModel) { { newValue ->
-        modVersionThreeViewModel.setBoxOneDoubleTap(newValue)
-    } }
-    val setBoxOneDragging:(newValue:Boolean)->Unit = remember(modVersionThreeViewModel) { { newValue ->
-        modVersionThreeViewModel.setBoxOneZIndex(newValue)
-    } }
-
-    /******BOX-Two FUNCTIONS*******/
-    val setBoxTwoDoubleTap:(newValue:Boolean)->Unit = remember(modVersionThreeViewModel) { { newValue ->
-        modVersionThreeViewModel.setBoxTwoDoubleTap(newValue)
-    } }
-
-    val setBoxTwoDragging:(newValue:Boolean)->Unit = remember(modVersionThreeViewModel) { { newValue ->
-        modVersionThreeViewModel.setBoxTwoZIndex(newValue)
-    } }
-
-    /******BOX-Three FUNCTIONS*******/
-
-    val setBoxThreeDoubleTap:(newValue:Boolean)->Unit = remember(modVersionThreeViewModel) { { newValue ->
-        modVersionThreeViewModel.setBoxThreeDoubleTap(newValue)
-    } }
-
-    val setBoxThreeDragging:(newValue:Boolean)->Unit = remember(modVersionThreeViewModel) { { newValue ->
-        modVersionThreeViewModel.setBoxThreeZIndex(newValue)
-    } }
-
-    /******Generic FUNCTIONS*******/
-
-
-    val setDoubleClickAndDragFalse:()->Unit = remember(modVersionThreeViewModel) { {
-
-        modVersionThreeViewModel.updateDoubleClickAndDrag(false)
-    } }
-
-    val manageAutoModMessage:(messageId:String,action:String)->Unit = remember(modVersionThreeViewModel) { {messageId,action->
-        modViewViewModel.manageAutoModMessage(messageId,action)
-
-    } }
-
-    val changeActualTextFieldValue:(String, TextRange) -> Unit = remember(streamViewModel) { { text, textRange ->
-        streamViewModel.changeActualTextFieldValue(text, textRange)
-    } }
-
-    val sendMessageToWebSocket:(String) -> Unit = remember(streamViewModel) { { message ->
-        streamViewModel.sendMessage(message)
-        streamViewModel.updateMostFrequentEmoteList()
-    } }
-    val setIndex:(Int)->Unit = remember(modVersionThreeViewModel) { { newValue ->
-
-        modVersionThreeViewModel.setIndex(newValue)
-    } }
-
-    val changeAutoModQueueChecked:(Boolean)->Unit = remember(modVersionThreeViewModel) { {newValue->
-        modViewViewModel.changeAutoModQueueChecked(newValue)
-
-    } }
-    val changeModActionsChecked:(Boolean)->Unit = remember(modVersionThreeViewModel) { {newValue->
-        modViewViewModel.changeModActionsChecked(newValue)
-
-    } }
-
-
-
-    val updateMostFrequentEmoteList:(EmoteNameUrl)->Unit =remember(streamViewModel) { {
-        streamViewModel.updateTemporaryMostFrequentList(it)
-    } }
-
-    /*******ALL FUNCTIONS RELATED TO CHAT SETTINGS SIZE***********/
-    val changeBadgeSize:(Float) -> Unit = remember(chatSettingsViewModel) { {newValue ->
-        chatSettingsViewModel.changeBadgeSize(newValue)
-    } }
-    val changeEmoteSize:(Float) -> Unit = remember(chatSettingsViewModel) { {newValue ->
-        chatSettingsViewModel.changeEmoteSize(newValue)
-    } }
-    val changeUsernameSize:(Float) -> Unit = remember(chatSettingsViewModel) { {newValue ->
-        chatSettingsViewModel.changeUsernameSize(newValue)
-    } }
-    val changeMessageSize:(Float) -> Unit = remember(chatSettingsViewModel) { {newValue ->
-        chatSettingsViewModel.changeMessageSize(newValue)
-    } }
-    val changeLineHeight:(Float) -> Unit = remember(chatSettingsViewModel) { {newValue ->
-        chatSettingsViewModel.changeLineHeight(newValue)
-    } }
-    val changeCustomUsernameColor:(Boolean) -> Unit = remember(chatSettingsViewModel) { {newValue ->
-        chatSettingsViewModel.changeCustomUsernameColor(newValue)
-    } }
-
-
-    val resolveUnbanRequest:(String,UnbanStatusFilter)->Unit = remember(modViewViewModel) { {id,status->
-        modViewViewModel.resolveUnbanRequest(id,status)
-
-    } }
-    val updateOptionalResolutionText:(String)->Unit = remember(modViewViewModel) { {newText->
-        modViewViewModel.updateOptionalResolutionText(newText)
-
-    } }
-
-    val changeSelectedFollowersModeItem:(ListTitleValue)->Unit = remember(modViewViewModel) { { newValue->
-        modViewViewModel.changeSelectedFollowersModeItem(newValue)
-
-    } }
-    val changeSelectedSlowModeItem:(ListTitleValue)->Unit = remember(modViewViewModel) { { newValue->
-        modViewViewModel.changeSelectedSlowModeItem(newValue)
-
-    } }
-    val updateSubscriberOnly:(Boolean)->Unit = remember(modViewViewModel) { { newValue->
-        modViewViewModel.updateSubscriberOnly(newValue)
-
-    } }
-    val sortUnbanRequestList:(String)->Unit = remember(modViewViewModel) { { status->
-        modViewViewModel.sortUnbanRequestList(status)
-
-    } }
-    val retryGetUnbanRequest:()->Unit = remember(modViewViewModel) { {
-        modViewViewModel.retryGetUnbanRequest()
-
-    } }
-    val changeUnbanRequestChecked:(Boolean)->Unit = remember(modViewViewModel) { { newValue->
-        modViewViewModel.changeUnbanRequestChecked(newValue)
-
-    } }
-
-    ModalBottomSheetLayout(
-        sheetState=unbanRequestModalState,
-        sheetContent = {
-            Box {
-                Column(modifier = Modifier
-                    .fillMaxWidth()
-                    .background(MaterialTheme.colorScheme.primary)
-                    .padding(10.dp)){
-                    NewContentBanner(
-                        clickedUsername = modViewViewModel.clickedUnbanRequestUser.value.userName,
-                        clickedMessage = modViewViewModel.clickedUnbanRequestUser.value.message,
-                        clickedUserInfo=modViewViewModel.clickedUnbanRequestInfo.value,
-                        clickedStatus = modViewViewModel.clickedUnbanRequestUser.value.status,
-                    )
-                    ClickedUserMessages(
-                        globalTwitchEmoteContentMap = chatSettingsViewModel.globalEmoteMap.value,
-                        channelTwitchEmoteContentMap = chatSettingsViewModel.inlineContentMapChannelEmoteList.value,
-                        globalBetterTTVEmoteContentMap = chatSettingsViewModel.betterTTVGlobalInlineContentMapChannelEmoteList.value,
-                        channelBetterTTVEmoteContentMap = chatSettingsViewModel.betterTTVChannelInlineContentMapChannelEmoteList.value,
-                        sharedBetterTTVEmoteContentMap = chatSettingsViewModel.betterTTVSharedInlineContentMapChannelEmoteList.value,
-                        clickedUsernameChatsWithDateSentImmutable = streamViewModel.clickedUsernameChatsDateSentImmutable.value,
-                        resolveUnbanRequest={id,status->
-                            resolveUnbanRequest(id,status)
-                        },
-                        clickedRequestId = modViewViewModel.clickedUnbanRequestUser.value.requestId,
-                        clickedStatus = modViewViewModel.clickedUnbanRequestUser.value.status,
-                        resolutionText=modViewViewModel.optionalResolutionText.value,
-                        updateResolutionText={newText->
-                           updateOptionalResolutionText(newText)
-                        }
-                    )
-
-                }
-                val response = modViewViewModel.resolveUnbanRequest.value
-                when(response){
-                    is Response.Loading ->{
-                        Spacer(
-                            modifier = Modifier
-                                .fillMaxSize()
-                                .disableClickAndRipple()
-                                .background(
-                                    color = Color.Black.copy(alpha = .7f)
-                                )
-                        )
-                        CircularProgressIndicator(modifier = Modifier.align(Alignment.Center))
-                    }
-                    is Response.Success ->{
-                        if(!response.data){
-                            Log.d("unbanrequestModalCheck","it should hide")
-
-                            Spacer(
-                                modifier = Modifier
-                                    .fillMaxSize()
-                                    .disableClickAndRipple()
-                                    .background(
-                                        color = Color.Black.copy(alpha = .7f)
-                                    )
-                            )
-                            Row(modifier = Modifier.align(Alignment.Center)){
-                                Icon(painter = painterResource(id =R.drawable.baseline_check_24), contentDescription = "successful request",tint = Color.Green)
-                               Text("Success",color = MaterialTheme.colorScheme.onPrimary)
-                                Icon(painter = painterResource(id =R.drawable.baseline_check_24), contentDescription = "successful request",tint = Color.Green)
-                            }
-                        }
-
-                    }
-                    is Response.Failure ->{
-                        Spacer(
-                            modifier = Modifier
-                                .fillMaxSize()
-                                .disableClickAndRipple()
-                                .background(
-                                    color = Color.Black.copy(alpha = .7f)
-                                )
-                        )
-                        Row(modifier = Modifier.align(Alignment.Center)){
-                            Icon(painter = painterResource(id =R.drawable.baseline_close_24), contentDescription = "Failed request",tint = Color.Red)
-                            Text("Failed",color = MaterialTheme.colorScheme.onPrimary)
-                            Icon(painter = painterResource(id =R.drawable.baseline_close_24), contentDescription = "Failed request",tint = Color.Red)
-                        }
-                    }
-
-                }
-
-            }
-
-        }
-    ) {
-
-         ModalBottomSheetLayout(
-             sheetState = chatSettingsModalState,
-             sheetContent = {
-                 ChatSettingsColumn(
-                     advancedChatSettings = streamViewModel.advancedChatSettingsState.value,
-                     changeAdvancedChatSettings = { newValue ->
-                         streamViewModel.updateAdvancedChatSettings(
-                             newValue
-                         )
-                     },
-                     changeNoChatMode = { newValue -> streamViewModel.setNoChatMode(newValue) },
-
-                     chatSettingsEnabled = streamViewModel.state.value.loggedInUserData?.mod
-                         ?: false,
-
-                     followerModeListImmutable = followerModeListImmutable,
-                     slowModeListImmutable = slowModeListImmutable,
-
-                     selectedFollowersModeItem = modViewViewModel.uiState.value.selectedFollowerMode,
-                     changeSelectedFollowersModeItem = { newValue ->
-                         changeSelectedFollowersModeItem(
-                             newValue
-                         )
-                     },
-
-                     selectedSlowModeItem = modViewViewModel.uiState.value.selectedSlowMode,
-                     changeSelectedSlowModeItem = { newValue ->
-
-                         changeSelectedSlowModeItem(
-                             newValue
-                         )
-                     },
-                     emoteOnly = modViewViewModel.uiState.value.emoteOnly,
-                     setEmoteOnly = { newValue -> modViewViewModel.updateEmoteOnly(newValue) },
-                     subscriberOnly = modViewViewModel.uiState.value.subscriberOnly,
-                     setSubscriberOnly = { newValue ->
-                         updateSubscriberOnly(
-                             newValue
-                         )
-                     },
-
-                     badgeSize = chatSettingsViewModel.badgeSize.value,
-                     changeBadgeSize = { newValue -> changeBadgeSize(newValue) },
-                     emoteSize = chatSettingsViewModel.emoteSize.value,
-                     changeEmoteSize = { newValue -> changeEmoteSize(newValue) },
-                     usernameSize = chatSettingsViewModel.usernameSize.value,
-                     changeUsernameSize = { newValue -> changeUsernameSize(newValue) },
-                     messageSize = chatSettingsViewModel.messageSize.value,
-                     changeMessageSize = { newValue -> changeMessageSize(newValue) },
-                     lineHeight = chatSettingsViewModel.lineHeight.value,
-                     changeLineHeight = { newValue -> changeLineHeight(newValue) },
-                     customUsernameColor = chatSettingsViewModel.customUsernameColor.value,
-                     changeCustomUsernameColor = { newValue -> changeCustomUsernameColor(newValue) },
-
-                     )
-             }
-         ) {
-
-             ModalBottomSheetLayout(
-                 sheetBackgroundColor = MaterialTheme.colorScheme.primary,
-                 sheetState = clickedChatterModalState,
-                 sheetContent = {
-
-                     TestingNewBottomModal(
-                         clickedUsername = streamViewModel.clickedUIState.value.clickedUsername,
-                         textFieldValue = streamViewModel.textFieldValue,
-                         closeBottomModal = {},
-                         banned = streamViewModel.clickedUIState.value.clickedUsernameBanned,
-                         unbanUser = {
-                             streamViewModel.unBanUser()
-                         },
-                         isMod = streamViewModel.state.value.loggedInUserData?.mod ?: false,
-                         openTimeoutDialog = { streamViewModel.openTimeoutDialog.value = true },
-                         openBanDialog = { streamViewModel.openBanDialog.value = true },
-                         openWarnDialog = { streamViewModel.changeOpenWarningDialog(true) },
-                         clickedUsernameChatsDateSentImmutable = streamViewModel.clickedUsernameChatsDateSentImmutable.value,
-                         badgeInlineContentMap = chatSettingsViewModel.globalChatBadgesMap.value,
-                         clickedUserBadgeList = streamViewModel.clickedUserBadgesImmutable.value,
-                         globalTwitchEmoteContentMap = chatSettingsViewModel.globalEmoteMap.value,
-                         channelTwitchEmoteContentMap = chatSettingsViewModel.inlineContentMapChannelEmoteList.value,
-                         globalBetterTTVEmoteContentMap = chatSettingsViewModel.betterTTVGlobalInlineContentMapChannelEmoteList.value,
-                         channelBetterTTVEmoteContentMap = chatSettingsViewModel.betterTTVChannelInlineContentMapChannelEmoteList.value,
-                         sharedBetterTTVEmoteContentMap = chatSettingsViewModel.betterTTVSharedInlineContentMapChannelEmoteList.value,
-                     )
-                 }
-             ) {
-                 ModVersionThree(
-                     boxOneOffsetY = modVersionThreeViewModel.boxOneOffsetY,
-                     setBoxOneOffset = { newValue ->
-                         modVersionThreeViewModel.setBoxOneOffset(
-                             newValue
-                         )
-                     },
-                     boxOneDragState = modVersionThreeViewModel.boxOneDragState,
-                     boxOneSection = modVersionThreeViewModel.boxOneSection,
-                     boxOneIndex = modVersionThreeViewModel.boxOneIndex,
-                     boxOneDragging = modVersionThreeViewModel.boxesZIndex.value.boxOneIndex,
-                     setBoxOneDragging = { newValue ->
-                         Log.d("LoggingTheDragging", "ONE")
-                         setBoxOneDragging(newValue)
-                     },
-                     setBoxOneIndex = { newValue ->
-                         modVersionThreeViewModel.syncBoxOneIndex(
-                             newValue
-                         )
-                     },
-                     deleteBoxOne = modVersionThreeViewModel.deleteBoxOne,
-                     boxOneHeight = modVersionThreeViewModel.boxOneHeight,
-                     boxOneDoubleTap = modVersionThreeViewModel.doubleTap.value.boxOneDoubleTap,
-                     setBoxOneDoubleTap = { newValue ->
-                         setBoxOneDoubleTap(newValue)
-                     },
-
-                     /*************** BOX TWO PARAMETERS***************************************************************/
-
-                     /*************** BOX TWO PARAMETERS***************************************************************/
-                     boxTwoOffsetY = modVersionThreeViewModel.boxTwoOffsetY,
-                     setBoxTwoOffset = { newValue ->
-                         modVersionThreeViewModel.setBoxTwoOffset(
-                             newValue
-                         )
-                     },
-                     boxTwoDragState = modVersionThreeViewModel.boxTwoDragState,
-                     boxTwoSection = modVersionThreeViewModel.boxTwoSection,
-                     boxTwoIndex = modVersionThreeViewModel.boxTwoIndex,
-                     boxTwoDragging = modVersionThreeViewModel.boxesZIndex.value.boxTwoIndex,
-                     setBoxTwoDragging = { newValue -> setBoxTwoDragging(newValue) },
-                     setBoxTwoIndex = { newValue ->
-                         Log.d("LoggingTheDragging", "TWO")
-                         modVersionThreeViewModel.syncBoxTwoIndex(newValue)
-                     },
-                     deleteBoxTwo = modVersionThreeViewModel.deleteBoxTwo,
-                     boxTwoHeight = modVersionThreeViewModel.boxTwoHeight,
-                     boxTwoDoubleTap = modVersionThreeViewModel.doubleTap.value.boxTwoDoubleTap,
-                     setBoxTwoDoubleTap = { newValue ->
-                         setBoxTwoDoubleTap(newValue)
-                     },
-
-                     /*************** BOX THREE PARAMETERS*****************************************************************/
-
-                     /*************** BOX THREE PARAMETERS*****************************************************************/
-                     boxThreeOffsetY = modVersionThreeViewModel.boxThreeOffsetY,
-                     setBoxThreeOffset = { newValue ->
-                         modVersionThreeViewModel.setBoxThreeOffset(
-                             newValue
-                         )
-                     },
-                     boxThreeDragState = modVersionThreeViewModel.boxThreeDragState,
-                     boxThreeSection = modVersionThreeViewModel.boxThreeSection,
-                     boxThreeIndex = modVersionThreeViewModel.boxThreeIndex,
-                     boxThreeDragging = modVersionThreeViewModel.boxesZIndex.value.boxThreeIndex,
-                     setBoxThreeDragging = { newValue ->
-                         setBoxThreeDragging(newValue)
-                     },
-                     setBoxThreeIndex = { newValue ->
-                         Log.d("LoggingTheDragging", "THREE")
-                         modVersionThreeViewModel.syncBoxThreeIndex(newValue)
-                     },
-                     deleteBoxThree = modVersionThreeViewModel.deleteBoxThree,
-                     boxThreeHeight = modVersionThreeViewModel.boxThreeHeight,
-                     boxThreeDoubleTap = modVersionThreeViewModel.doubleTap.value.boxThreeDoubleTap,
-                     setBoxThreeDoubleTap = { newValue ->
-                         setBoxThreeDoubleTap(newValue)
-
-                     },
-
-
-                     /*************** GENERICS PARAMETERS*****************************************************************/
-
-
-                     /*************** GENERICS PARAMETERS*****************************************************************/
-                     updateIndex = { newValue ->
-                         setIndex(newValue) //todo: wrap this in a remember function
-                     },
-                     showError = modVersionThreeViewModel.showPlacementError.value,
-                     sectionTwoHeight = modVersionThreeViewModel.section2height,
-                     sectionThreeHeight = modVersionThreeViewModel.section3Height,
-                     closeModView = { closeModView() },
-                     fullChatMode = modVersionThreeViewModel.fullChat.value,
-                     deleteOffset = modVersionThreeViewModel.deleteOffset,
-
-                     smallChat = { setDragging ->
-                         SmallChat(
-                             twitchUserChat = twitchUserChat,
-                             showBottomModal = {
-                                 scope.launch { clickedChatterModalState.show() }
-                             },
-                             updateClickedUser = { username, userId, banned, isMod ->
-                                 updateClickedUser( //todo: I need to add this
-                                     username,
-                                     userId,
-                                     banned,
-                                     isMod
-                                 )
-                             },
-                             showTimeoutDialog = {
-                                 streamViewModel.openTimeoutDialog.value = true
-                             },
-                             showBanDialog = { streamViewModel.openBanDialog.value = true },
-                             doubleClickMessage = { username ->
-                                 doubleClickChat(username)
-                             },
-                             deleteChatMessage = { messageId ->
-                                 streamViewModel.deleteChatMessage(
-                                     messageId
-                                 )
-                             },
-                             //todo:change back to --> true for testing
-                             isMod = streamViewModel.state.value.loggedInUserData?.mod ?: false,
-                             setDragging = { value ->
-                                 Log.d("DOUBLECLICKDRAGGING", "click from the outside")
-                                 setDragging()
-                             },
-                             doubleClickAndDrag = modVersionThreeViewModel.doubleClickAndDrag.value,
-                             setDoubleClickAndDragFalse = {
-                                 setDoubleClickAndDragFalse()
-                             },
-                             badgeListMap = streamViewModel.badgeListMap.value,
-                             usernameSize = chatSettingsViewModel.usernameSize.value,
-                             messageSize = chatSettingsViewModel.messageSize.value,
-                             lineHeight = chatSettingsViewModel.lineHeight.value,
-                             useCustomUsernameColors = chatSettingsViewModel.customUsernameColor.value,
-                             globalTwitchEmoteContentMap = chatSettingsViewModel.globalEmoteMap.value,
-                             channelTwitchEmoteContentMap = chatSettingsViewModel.inlineContentMapChannelEmoteList.value,
-                             globalBetterTTVEmoteContentMap = chatSettingsViewModel.betterTTVGlobalInlineContentMapChannelEmoteList.value,
-                             channelBetterTTVEmoteContentMap = chatSettingsViewModel.betterTTVChannelInlineContentMapChannelEmoteList.value,
-                             sharedBetterTTVEmoteContentMap = chatSettingsViewModel.betterTTVSharedInlineContentMapChannelEmoteList.value,
-                         )
-                     },
-                     fullChat = { setDragging ->
-                         FullChatModView(
-                             twitchUserChat = twitchUserChat,
-                             showBottomModal = {
-                                 scope.launch { clickedChatterModalState.show() }
-                             },
-                             updateClickedUser = { username, userId, banned, isMod ->
-                                 updateClickedUser(
-                                     username,
-                                     userId,
-                                     banned,
-                                     isMod
-                                 )
-                             },
-                             showTimeoutDialog = {
-                                 streamViewModel.openTimeoutDialog.value = true
-                             },
-                             showBanDialog = { streamViewModel.openBanDialog.value = true },
-                             doubleClickMessage = { username ->
-                                 doubleClickChat(username)
-
-                             },
-
-                             showOuterBottomModalState = {
-                                 Log.d("BottomModalClicked", "showOuterBottomModalState Clicked")
-                                 //seems fine that it is empy
-//                    scope.launch {
-//
-//                    }
-                             },
-                             newFilterMethod = { newTextValue ->
-                                 streamViewModel.newParsingAgain(
-                                     newTextValue
-                                 )
-                             },
-
-                             orientationIsVertical = true,
-
-                             //todo:change back to --> streamViewModel.state.value.loggedInUserData?.mod ?: false
-                             isMod = true,
-                             clickedAutoCompleteText = { username ->
-                                 streamViewModel.autoTextChange(username)
-                             },
-                             showModal = {
-                                 //todo: This is what is clicked when I want to launch the bottom modal
-                                 scope.launch {
-                                     chatSettingsModalState.show()
-                                     // clickedChatterModalState.show()
-                                 }
-                             },
-                             notificationAmount = 0,
-                             textFieldValue = streamViewModel.textFieldValue,
-                             sendMessageToWebSocket = { message ->
-                                 sendMessageToWebSocket(message)
-                             },
-                             noChat = streamViewModel.advancedChatSettingsState.value.noChatMode,
-                             deleteChatMessage = { messageId ->
-                                 streamViewModel.deleteChatMessage(
-                                     messageId
-                                 )
-                             },
-                             clickedCommandAutoCompleteText = { clickedValue ->
-                                 streamViewModel.clickedCommandAutoCompleteText(
-                                     clickedValue
-                                 )
-                             },
-                             hideSoftKeyboard = {
-                                 hideSoftKeyboard()
-
-                             },
-                             emoteBoardGlobalList = streamViewModel.globalEmoteUrlList.value,
-                             updateTextWithEmote = { newValue ->
-                                 streamViewModel.addEmoteToText(
-                                     newValue
-                                 )
-                             },
-                             emoteBoardChannelList = streamViewModel.channelEmoteUrlList.value,
-                             deleteEmote = { streamViewModel.deleteEmote() },
-                             showModView = {
-                                 closeModView()
-                             },
-                             fullMode = modVersionThreeViewModel.fullChat.value,
-                             setDragging = {
-                                 Log.d("doubleClickingThings", "CLICKED")
-                                 setDragging()
-                             },
-                             emoteBoardMostFrequentList = streamViewModel.mostFrequentEmoteListTesting.value,
-                             updateMostFrequentEmoteList = { value ->
-                                 updateMostFrequentEmoteList(value)
-                             },
-                             globalBetterTTVEmotes = streamViewModel.globalBetterTTVEmotes.value,
-                             channelBetterTTVResponse = streamViewModel.channelBetterTTVEmote.value,
-                             sharedBetterTTVResponse = streamViewModel.sharedChannelBetterTTVEmote.value,
-                             userIsSub = streamViewModel.state.value.loggedInUserData?.sub ?: false,
-                             forwardSlashes = streamViewModel.forwardSlashCommandImmutable.value,
-                             filteredChatListImmutable = streamViewModel.filteredChatListImmutable.value,
-                             actualTextFieldValue = streamViewModel.textFieldValue.value,
-                             changeActualTextFieldValue = { text, textRange ->
-                                 changeActualTextFieldValue(text, textRange)
-                             },
-                             badgeListMap = streamViewModel.badgeListMap.value,
-                             usernameSize = chatSettingsViewModel.usernameSize.value,
-                             messageSize = chatSettingsViewModel.messageSize.value,
-                             lineHeight = chatSettingsViewModel.lineHeight.value,
-                             useCustomUsernameColors = chatSettingsViewModel.customUsernameColor.value,
-                             globalTwitchEmoteContentMap = chatSettingsViewModel.globalEmoteMap.value,
-                             channelTwitchEmoteContentMap = chatSettingsViewModel.inlineContentMapChannelEmoteList.value,
-                             globalBetterTTVEmoteContentMap = chatSettingsViewModel.betterTTVGlobalInlineContentMapChannelEmoteList.value,
-                             channelBetterTTVEmoteContentMap = chatSettingsViewModel.betterTTVChannelInlineContentMapChannelEmoteList.value,
-                             sharedBetterTTVEmoteContentMap = chatSettingsViewModel.betterTTVSharedInlineContentMapChannelEmoteList.value,
-                             lowPowerMode = streamViewModel.lowPowerModeActive.value,
-                             channelName = streamViewModel.channelName.value ?: "",
-                             showChannelInformationModal = {  }
-
-                         )
-
-                     },
-                     modActionsList = modViewViewModel.modActionsList,
-                     modActionStatus = modViewViewModel.modViewStatus.value.modActions,
-                     autoModMessageList = modViewViewModel.autoModMessageList,
-                     autoModStatus = modViewViewModel.modViewStatus.value.autoModMessageStatus,
-                     manageAutoModMessage = { messageId, action ->
-                         manageAutoModMessage(messageId, action)
-                     },
-                     changeAutoModQueueChecked = { value ->
-                         changeAutoModQueueChecked(value)    //todo: wrap this in a remember function
-                     },
-                     changeModActionsChecked = { value ->
-                         changeModActionsChecked(value) //todo: wrap this in a remember function
-                     },
-                     autoModQueueChecked = modViewViewModel.uiState.value.autoModMessagesNotifications,
-                     modActionsChecked = modViewViewModel.uiState.value.modActionNotifications,
-                     doubleClickAndDrag = modVersionThreeViewModel.doubleClickAndDrag.value,
-                     setDoubleClickAndDragFalse = {
-                         setDoubleClickAndDragFalse()
-                     },
-                     autoModMessageListImmutableCollection = modViewViewModel.autoModMessageListImmutable.value,
-                     modActionListImmutableCollection = modViewViewModel.modActionListImmutable.value,
-                     unbanRequestResponse = modViewViewModel.unbanRequestResponse.value,
-                     showUnbanRequestModal={showUnbanRequestBottomModal()},
-                     updateClickedUnbanRequest={ username,text,userId,requestId,status ->
-                         updateClickedUser(username,userId,false,false)
-                         modViewViewModel.updateClickedUnbanRequestUser(username,text,userId,requestId,status)
-                     },
-                     immutableUnbanRequestList = modViewViewModel.getUnbanRequestList.value,
-                     sortUnbanRequest={status ->
-                        sortUnbanRequestList(status)
-                                      },
-                     retryUnbanRequests={
-                             retryGetUnbanRequest()
-                                        },
-                     unbanRequestChecked = modViewViewModel.uiState.value.unbanRequestNotifications,
-                     changeUnbanRequestChecked = {value ->
-                         //todo: WRAP THIS----------------------------------------
-                       changeUnbanRequestChecked(value)
-                     }
-
-
-
-                     )
-             }
-         }
-
-     }
-
-}
 
 /**
  * - restartable
@@ -863,7 +196,7 @@ fun ModVersionThree(
     //immutable Lists
     autoModMessageListImmutableCollection: AutoModMessageListImmutableCollection,
     modActionListImmutableCollection: ModActionListImmutableCollection,
-    immutableUnbanRequestList:UnbanRequestItemImmutableCollection,
+    immutableUnbanRequestList: UnbanRequestItemImmutableCollection,
 
     unbanRequestResponse: UnAuthorizedResponse<List<UnbanRequestItem>>,
     showUnbanRequestModal:()->Unit,
@@ -1323,7 +656,7 @@ fun UnbanRequests(
     unbanRequestResponse: UnAuthorizedResponse<List<UnbanRequestItem>>,
     showUnbanRequestModal:()->Unit,
     updateClickedUnbanRequest: (String, String, String,String,String) -> Unit,
-    immutableUnbanRequestList:UnbanRequestItemImmutableCollection,
+    immutableUnbanRequestList: UnbanRequestItemImmutableCollection,
     sortUnbanRequest:(String)->Unit,
     retryUnbanRequests:()->Unit,
 ){
@@ -1380,10 +713,10 @@ fun UnbanRequestLazyColumn(
     doubleClickAndDrag: Boolean,
     showUnbanRequestModal:()->Unit,
     updateClickedUnbanRequest: (String, String, String,String,String) -> Unit,
-    immutableUnbanRequestList:UnbanRequestItemImmutableCollection,
+    immutableUnbanRequestList: UnbanRequestItemImmutableCollection,
     sortUnbanRequest:(String)->Unit,
 
-){
+    ){
     LazyColumn(
         modifier = Modifier
             .fillMaxSize()
