+++ app/src/main/java/com/example/clicker/presentation/moderatedChannelsHome/ModChannelView.kt
@@ -14,7 +14,7 @@ import com.example.clicker.presentation.home.HomeViewModel
 import com.example.clicker.presentation.authentication.logout.LogoutViewModel
 import com.example.clicker.presentation.moderatedChannelsHome.views.MainModView
 import com.example.clicker.presentation.moderatedChannelsHome.views.ModChannelsBottomModalSheetContent
-import com.example.clicker.presentation.modView.ModViewViewModel
+import com.example.clicker.presentation.enhancedModView.viewModels.ModViewViewModel
 import com.example.clicker.presentation.search.SearchViewModel
 import com.example.clicker.presentation.stream.AutoModViewModel
 import com.example.clicker.presentation.stream.StreamViewModel
