+++ app/src/test/java/com/example/clicker/network/repository/util/EmoteParsingTest.kt
@@ -1,10 +1,5 @@
 package com.example.clicker.network.repository.util
 
-import androidx.compose.foundation.text.InlineTextContent
-import com.example.clicker.presentation.modView.BoxTypeIndex
-import org.junit.Assert
-import org.junit.Test
-
 class EmoteParsingTest {
 
 }
